1. Re-org the folder structure. 2. Add WoofTextService as another sample 3. Add IMServiceMgr to enumerate installed input method for client/server usage. 4. Add 2 attributes in JSON for IMServiceMgr instantiating service instance.
